requirements.txt
aiogram~=3.15.0
aiosqlite~=0.20.0
fastapi[all]~=0.115.5
greenlet~=3.1.1
mistralai~=1.2.3
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36
pydantic~=2.9.2
requests~=2.32.3